requirements.txt
torch==2.8.0
torchvision
transformers
datasets
hf_xet
sentencepiece
faker
Pillow
opencv-python
albumentations
tqdm
numpy
matplotlib
scikit-learn
pytest
pytest-cov
flake8
black
isort
wandb
protobuf
Levenshtein
